Reorganize feature additions
